Added a basic React app w/ upload.
This includes adding a React app as well as a common Form component with styles
and CSRF tooling, an HTTP library, a common padded page element and styles. This
also includes setting up API controllers and putting in a simple version of the
upload API controller. Lastly, this includes a file upload page which includes
reading the file and sending it to the back end via the API.